Moved logic for updating the Activity Registry
