+++ src/modules/Elsa.Workflows.Core/Features/WorkflowsFeature.cs
@@ -137,6 +137,7 @@ public class WorkflowsFeature : FeatureBase
             .AddSingleton<IActivityDescriber, ActivityDescriber>()
             .AddSingleton<IActivityRegistry, ActivityRegistry>()
             .AddScoped<IActivityRegistryLookupService, ActivityRegistryLookupService>()
+            .AddScoped<IActivityRegistryUpdateService, ActivityRegistryUpdateService>()
             .AddSingleton<IPropertyDefaultValueResolver, PropertyDefaultValueResolver>()
             .AddSingleton<IPropertyUIHandlerResolver, PropertyUIHandlerResolver>()
             .AddSingleton<IActivityFactory, ActivityFactory>()
